requirements.txt
# AI Voice Assistant Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Audio Processing
numpy>=1.21.0
scipy==1.11.4
soundfile==0.12.1

# Speech Recognition (Whisper)
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# Voice Activity Detection
silero-vad==4.0.0

# Text-to-Speech
edge-tts==7.2.0
pyttsx3==2.99
comtypes==1.4.11  # Windows TTS support
pywin32==311      # Windows TTS support

# RAG & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM Integration
openai==1.3.7
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pathlib==1.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
